Crate concordium_base

A library that defines common types and functionality that is needed by various Rust projects. The scope of this library is meant to be limited to core chain definitions. At present this is an internal library with an unstable API. Where necessary the functionality should be re-exported, for example like it is in the Rust-SDK.

This library should always be possible to compile for android, iOS, Wasm, and x86 code. Some parts may be feature gated to work around platform specific limitations though.

This library also exports other core crypto dependencies so that consumers may simplify their dependencies. Users are intended to get the re-exported dependencies through the library, instead of separately.

Re-exports

pub use aggregate_sig;

pub use concordium_contracts_common as contracts_common;

pub use crypto_common as common;

pub use ecvrf;

pub use eddsa_ed25519;

pub use encrypted_transfers;

pub use id;

pub use random_oracle;

Modules

base

Basis type definitions that are used throughout the crate.

constants

hashes

Different types of hashes based on SHA256.

smart_contracts

transactions

Definition of transactions and other transaction-like messages, together with their serialization, signing, and similar auxiliary methods.

updates

Definitions and functionality related to chain updates.