Module concordium_base::hashes

Different types of hashes based on SHA256.

Structs

HashBytes

A general wrapper around Sha256 hash. This is used to add type safety to a hash that is used in different context. The underlying value is always the same, but the phantom type variable makes it impossible to mistakenly misuse the hashes.

IncorrectLength

Enums

ElectionNonceMarker

Used as a phantom type to indicate a hash is a hash is a leadership election nonce.

HashFromStrError

Possible errors when converting from a string to any kind of hash. String representation of hashes is as base 16.

PureHashMarker

Used as a phantom type to indicate a hash is a hash with no specific meaning.

StateMarker

Used as a phantom type to indicate a hash is a block state hash.

TransactionSignMarker

Used as a phantom type to indicate a hash is a transaction sign hash, i.e., the hash that is signed.

UpdateSignMarker

Used as a phantom type to indicate a hash is an update sign hash, i.e., the hash that is signed to make an update instruction.

Type Definitions

BlockHash

Hash of a block.

Hash

A Sha256 with no specific meaning.

LeadershipElectionNonce

The leadership election nonce is an unpredictable value updated once an epoch to make sure that bakers cannot predict too far in the future when they will win blocks.

StateHash

Hash of the block state that is included in a block.

TransactionHash

Hash of a transaction.

TransactionSignHash

Hash that is signed by the account holder’s keys to make a transaction signature.

UpdateSignHash

Hash that is signed by the governance keys to make an update instruction signature.