Type Definition concordium_base::hashes::LeadershipElectionNonce

pub type LeadershipElectionNonce = HashBytes<ElectionNonceMarker>;

The leadership election nonce is an unpredictable value updated once an epoch to make sure that bakers cannot predict too far in the future when they will win blocks.